+++ src/db/schema/mod.rs
@@ -19,6 +19,7 @@ mod v26_to_v27;
 mod v27_to_v28;
 mod v28_to_v29;
 mod v29_to_v30;
+mod v30_to_v31;
 mod v2_to_v3;
 mod v3_to_v4;
 mod v4_to_v5;
@@ -57,6 +58,7 @@ use v26_to_v27::UPGRADE_26_TO_27_PRE_SQL;
 use v27_to_v28::migrate_v27_to_v28;
 use v28_to_v29::{migrate_v28_to_v29, UPGRADE_28_TO_29_PRE_SQL};
 use v29_to_v30::{migrate_v29_to_v30, UPGRADE_29_TO_30_PRE_SQL};
+use v30_to_v31::{migrate_v30_to_v31, UPGRADE_30_TO_31_PRE_SQL};
 use v2_to_v3::UPGRADE_2_TO_3_SQL;
 use v3_to_v4::UPGRADE_3_TO_4_SQL;
 use v4_to_v5::UPGRADE_4_TO_5_SQL;
@@ -173,5 +175,10 @@ pub const MIGRATION_28_TO_29: Migration = Migration::Transacted {
 pub const MIGRATION_29_TO_30: Migration = Migration::Transacted {
     pre_sql: Some(UPGRADE_29_TO_30_PRE_SQL),
     code_fn: Some(migrate_v29_to_v30),
+    post_sql: None,
+};
+pub const MIGRATION_30_TO_31: Migration = Migration::Transacted {
+    pre_sql: Some(UPGRADE_30_TO_31_PRE_SQL),
+    code_fn: Some(migrate_v30_to_v31),
     post_sql: None,
 };